% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_14_data.r
\docType{data}
\name{cohort_ef}
\alias{cohort_ef}
\title{Simulated cohort of subjects exposed to ct scans}
\format{A data frame with 93017 rows and 8 variables:
\describe{
  \item{id}{subject identificator, integer}
  \item{sex}{sex of the subject, 1-male / 2-female}
  \item{entry_age}{age of entry at the cohort, numeric}
  \item{exit_age}{age at exit of the cohort, numeric}
  \item{outcome}{disease or not, 0-no disease / 1-disease}
  \item{age}{age at scan, numeric}
  \item{dose}{dose of the scan, numeric in mGy}
  \item{country}{country, character}
}}
\usage{
cohort_ef
}
\description{
List of scans recived by a cohort of subjects
}
\keyword{datasets}
