% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_12_ci_rERR.r
\name{confint.rERR}
\alias{confint.rERR}
\title{Confidence intervals}
\usage{
\method{confint}{rERR}(object)
}
\arguments{
\item{object}{an object of class rERR}
}
\value{
a list with the confidence intervals
}
\description{
Show the confidence intervals for each parameter of the model.
The likelihood ratio test ci for linear varaibles, and the Wald ci for the loglinear terms
}
\examples{
\\donotrun{ ci.rERR(fit)}
}
