\name{Dosimetry}
\alias{Dose}
\alias{annualDoseToCumulativeDose}
\alias{cumulativeDoseToAnnualDose}
\alias{hsumaAnnualDose}
\alias{hsumaCumulativeDose}
%% \alias{msiERR}
\alias{fitERR}
%% \alias{varERR}
\alias{inferERR}

\title{
Functions related to Excessive Relative Risk (ERR) model fitting and inference
}

\description{
  TODO
  \name{annualDoseToCumulativeDose}
  
  \name{cumulativeDoseToAnnualDose}

  \name{hsumaAnnualDose}

  \name{hsumaCumulativeDose}
}

\usage{
  TODO
}

\arguments{
  TODO
}

\details{
  TODO
}
\references{
  TODO
}
\keyword{TODO}
\seealso{
  TODO
%% Optional links to other man pages, e.g.
%% \code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
  %% TODO

}
