% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_07_exclusion.r
\name{f_exclusion}
\alias{f_exclusion}
\title{Exclusion period related with the latency - lag period}
\usage{
f_exclusion(formula, data, lag)
}
\arguments{
\item{formula}{Surv(entry_time,exit_time,outcome)~loglin(loglin_var1,..,loglin_varn)+\cr
lin(lin_var1,..,lin_varm)+strata(strat_var1,...strat_varp)}

\item{data}{initial data set}

\item{lag}{latency period}
}
\value{
a data set with the exclusion updated
}
\description{
Exclude subjects in the cohort that have less than lag time of follow up
}
\examples{
\dontrun{f_exclusion(formula,data,lag)}
}
