% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_04_lin_ERR_all.r
\name{f_fit_linERR_all}
\alias{f_fit_linERR_all}
\title{fit Excess Relative Risk Model}
\usage{
f_fit_linERR_all(formula, data, id_name, time_name, lag)
}
\arguments{
\item{formula}{Surv(entry_time,exit_time,outcome)~loglin(loglin_var1,..,loglin_varn)+lin(lin_var1,..,lin_varm)+strata(strat_var1,...strat_varp)}

\item{data}{data set returned from f_to_model_data}

\item{id_name}{name of variable containing the names of subjects}

\item{time_name}{name of the time variable}

\item{lag}{latency period}
}
\value{
rERR object with the estimation
}
\description{
function that calls the optimization (mle from stats4 package, so use optim), and return a rERR object with the estimation and summary
}
\examples{
\\donotrun{ f_fit_linERR_all(formula,data,id_name,time_name,lag)}
}
