% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_06_lrt.r
\name{f_lrt}
\alias{f_lrt}
\title{likelihood ratio test}
\usage{
f_lrt(fit1, fit2)
}
\arguments{
\item{fit1}{the nested model}

\item{fit2}{the nesting model}
}
\value{
a list containing the lrt statistic and the corresponding p_value from the Chi-square test
}
\description{
function that thtat compute the lrt test for a nested and a nesting models
}
\examples{
\dontrun{ lrt(fit1,fit2)}
}
