% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_05_parse_formula.r
\name{f_parse_formula}
\alias{f_parse_formula}
\title{Parse formula (internal use)}
\usage{
f_parse_formula(formula)
}
\arguments{
\item{formula}{Surv(entry_time,exit_time,outcome)~loglin(loglin_var1,..,loglin_varn)+\cr
lin(lin_var1,..,lin_varm)+strata(strat_var1,...strat_varp)}
}
\value{
list of terms in the formula
}
\description{
Return list with the terms and elements of the formula
}
\examples{
\dontrun{ f_parse_formula(formula)}
}
