% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_14_plot_fitERR.r
\name{f_plot_linERR}
\alias{f_plot_linERR}
\title{plot the likelihood}
\usage{
f_plot_linERR(object, formula, data, rsets, n_lin_vars, n_loglin_vars, id_name,
  time_name)
}
\arguments{
\item{object}{An rERR class object}

\item{formula}{Surv(entry_time,exit_time,outcome)~loglin(loglin_var1,..,loglin_varn)+\cr
lin(lin_var1,..,lin_varm)+strata(strat_var1,...strat_varp)}

\item{data}{data set returned from f_to_model_data}

\item{rsets}{list of risksets, output of f_risksets}

\item{n_lin_vars}{number of linear variables (attribute of the to_model_data)}

\item{n_loglin_vars}{number of loglinear varaibles (attribute of the to_model_data)}

\item{id_name}{name of variable containing the names of subjects}

\item{time_name}{name of the time variable}
}
\value{
rERR object with the estimation
}
\description{
plot the partial log likelihood function in the case of one dimension in the linear part
}
\examples{
\dontrun{f_fit_linERR(formula,data,rsets,n_lin_vars,n_loglin_vars,id_name,time_name)}
}
