% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_00_to_event_table_ef_v2.r
\name{f_to_event_table_ef_v2}
\alias{f_to_event_table_ef_v2}
\title{Data transformation: Event format -> Event format required for the model (internal use)}
\usage{
f_to_event_table_ef_v2(id, start, stop, outcome, data, times, doses, covars)
}
\arguments{
\item{id}{name of variable containing the names of subjects}

\item{start}{name of the variable containing the start time}

\item{stop}{name of the variable containing the stop time}

\item{outcome}{name of the variable containging the outcome}

\item{data}{input data set}

\item{times}{name of the time variable}

\item{doses}{name of the dose variable}

\item{covars}{names of the covars required later in the model}
}
\value{
The data set with the event-row format, including the event of exit of the cohort
}
\description{
This function organize an input data set ef, same as f_to_event_table_ef_all but with start,stop and outcome from the formula
}
