% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_03_to_model_data.r
\name{f_to_model_data}
\alias{f_to_model_data}
\title{Data transformation: keep model variables and expand categorical variables (internal use)}
\usage{
f_to_model_data(formula, data, id_name, time_name)
}
\arguments{
\item{formula}{Surv(entry_time,exit_time,outcome)~loglin(loglin_var1,..,loglin_varn)+lin(lin_var1,..,lin_varm)+strata(strat_var1,...strat_varp)}

\item{data}{data set}

\item{id_name}{name of variable containing the names of subjects}

\item{time_name}{name of the time variable}
}
\value{
data set described below
}
\description{
Transform the data set in a closed form n_row|id_name|n_pe|entry_name|exit_name|outcome|time|linear_covariates|loglinear_covariates.
Expand if a variable is categorical to pure logical n_categories variables (excluding the reference category)
}
\examples{
\\donotrun{ f_to_model_data(formula,data,id_name='patientids',time_name='time')}
}
