% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_13_print_ci_rERR.r
\name{print.ci.rERR}
\alias{print.ci.rERR}
\title{Print ci rERR}
\usage{
\method{print}{ci.rERR}(x, ...)
}
\arguments{
\item{x}{a ci.ERR object}
}
\description{
Print the cofnidence intervals of rERR fit
}
