% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_11_print_summary_rERR.r
\name{print.summary.rERR}
\alias{print.summary.rERR}
\title{Print summary rERR}
\usage{
\method{print}{summary.rERR}(x, ...)
}
\arguments{
\item{x}{a summary.ERR object}

\item{\ldots}{for future methods}
}
\description{
Print the summary of rERR fit
}
