\name{Rerr-package}
\alias{Rerr-package}
\alias{Rerr}
\docType{package}
\title{
Excessive Relative Risk (ERR) model fitting and inference
}
\description{
\packageDescription{Rerr}
}
\details{
  TODO
%% The DESCRIPTION file:
%% \packageDESCRIPTION{Rerr}
%% \packageIndices{Rerr}
%% ~~ An overview of how to use the package, including the most important ~~
%% ~~ functions ~~
}
\author{
Zhuo Zhang, Daniel Stram

Maintainer: Zhuo Zhang 
}
\references{
  TODO
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%% ~~ the R documentation directory ~~
\keyword{ Rerr }
\seealso{
  TODO
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
  %% TODO
%% ~~ simple examples of the most important functions ~~
}
