% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_10_summary_rERR.r
\name{summary.rERR}
\alias{summary.rERR}
\title{summary of a ERR fit}
\usage{
\method{summary}{rERR}(object, ...)
}
\arguments{
\item{object}{an object of class rERR}

\item{\ldots}{for future methods}
}
\value{
a list with the summary elements
}
\description{
display summary of the parameter and statistics of the model
}
\examples{
\dontrun{summary(fit)}
}
